+++ misc/utils.py
@@ -11,8 +11,58 @@ import shutil
 import subprocess
 import sys
 
+from misc.deprecated_decorator import deprecated
 import misc.shell_utils as shell_utils
 
+@deprecated("[utils.check_is_fitted] Please use the version in `validation_utils`")
+def check_is_fitted(estimator, attributes, msg=None, all_or_any=all):
+    """Perform is_fitted validation for estimator.
+
+    N.B. This is essentially the same as the function in
+    `sklearn.utils.validation`; however, it *does not* check that `estimator`
+    has a `fit` method.
+
+    Checks if the estimator is fitted by verifying the presence of
+    "all_or_any" of the passed attributes and raises a NotFittedError with the
+    given message.
+    Parameters
+    ----------
+    estimator : estimator instance.
+        estimator instance for which the check is performed.
+    attributes : attribute name(s) given as string or a list/tuple of strings
+        Eg.:
+            ``["coef_", "estimator_", ...], "coef_"``
+
+    msg : string
+        The default error message is, "This %(name)s instance is not fitted
+        yet. Call 'fit' with appropriate arguments before using this method."
+        For custom messages if "%(name)s" is present in the message string,
+        it is substituted for the estimator name.
+        Eg. : "Estimator, %(name)s, must be fitted before sparsifying".
+
+    all_or_any : callable, {all, any}, default all
+        Specify whether all or any of the given attributes must exist.
+
+    Returns
+    -------
+    None
+
+    Raises
+    ------
+    NotFittedError
+        If the attributes are not found.
+    """
+    import sklearn.exceptions
+    
+    if msg is None:
+        msg = ("This %(name)s instance is not fitted yet. Call the appropriate "
+               "initialization, fit, etc., methods before using this method.")
+
+    if not isinstance(attributes, (list, tuple)):
+        attributes = [attributes]
+
+    if not all_or_any([hasattr(estimator, attr) for attr in attributes]):
+        raise sklearn.exceptions.NotFittedError(msg % {'name': type(estimator).__name__})
 
 def raise_deprecation_warning(function, new_module, final_version=None,
         old_module="misc"):
@@ -959,6 +1009,7 @@ def merge_sets(*set_args):
     ret = {item for s in set_args for item in s}
     return ret
 
+@deprecated("[utils.merge_dicts]: please use `toolz.dicttolz.merge` instead")
 def merge_dicts(*dict_args):
     """ Given any number of dicts, shallow copy and merge into a new dict,
     precedence goes to key value pairs in latter dicts.
@@ -1124,6 +1175,7 @@ def list_insert_list(l, to_insert, index):
     ret[index:index] = list(to_insert)
     return ret
 
+@deprecated("[utils.remove_keys]: please use `toolz.dicttolz.dissoc` instead")
 def remove_keys(d, to_remove):
     """ This function removes the given keys from the dictionary d. N.B.,
         "not in" is used to match the keys.