+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this repo will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/), 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.9] - 2018-04-18
+### Added
+- Helper to estimate categorical variable MLEs
+- Utilities for working with the SCIP solver
+- Helper to check if class attributes have been initialized similar to
+    `check_is_fitted` in sklearn
+- Standard validation helpers (`validation_utils`)
+- Helper to calculate many regression, multi-class classification metrics
+
 ## [0.2.8] - 2018-03-12
 ### Updated
 - Cross-validation helper to work with unsupervised learning